Drive `wendy os install` for Thor through the Go flash engine
Replace the Python-bootburn install path with a shared, cross-platform flow. The
common logic (flashpack plan/download, briefing, the download -> stage-1 -> stage-2
step list, and error hints) lives in one place; only stage-1 RCM boot, recovery
enumeration, and the stage-2 ADB transport differ per platform, provided by
per-OS hooks:

- Windows: the WinUSB backend (driver install on first use, RCM boot, and
  ADB-over-WinUSB), pinned to the chosen board by physical location.
- macOS/Linux: the existing gousb rcm/bringup + serverless adb client.

Stage-2 is the shared flashengine on all three. A stage-2 failure where the
gadget never re-enumerates (nothing written) shows the calm "gadget unreachable"
hint instead of the destructive bad-state recovery box. Also drops the shim
dispatch from the wendy entrypoint.
