+++ go/internal/cli/commands/os_install_thor_hw_darwin_linux.go
@@ -1,0 +1,140 @@
+//go:build darwin || linux
+
+package commands
+
+// macOS/Linux USB hooks for the shared Thor install flow: recovery-device
+// enumeration and stage-1 RCM boot use gousb (packages rcm/bringup); the stage-2
+// ADB transport is the serverless gousb adb client, which drives the shared
+// flashengine — no more python bootburn, adb shim, or PyYAML.
+
+import (
+	"errors"
+	"fmt"
+	"io"
+	"os"
+	"time"
+
+	"golang.org/x/sys/unix"
+
+	"github.com/wendylabsinc/wendy/go/internal/cli/tegraflash/adb"
+	"github.com/wendylabsinc/wendy/go/internal/cli/tegraflash/bringup"
+	"github.com/wendylabsinc/wendy/go/internal/cli/tegraflash/flashengine"
+	"github.com/wendylabsinc/wendy/go/internal/cli/tegraflash/flashpack"
+	"github.com/wendylabsinc/wendy/go/internal/cli/tegraflash/rcm"
+	"github.com/wendylabsinc/wendy/go/internal/cli/tui"
+)
+
+// thorPrepareHost stops a conflicting adb server that would claim the gadget.
+func thorPrepareHost(out io.Writer) error {
+	if stopConflictingADBServer() {
+		fmt.Fprintln(out, tui.InfoMessage("Stopped a running adb server (it would hold the Thor's flashing gadget)."))
+	}
+	return nil
+}
+
+// pickThorRecoveryDevice lists Jetsons in recovery mode and selects one, with a
+// rescan loop and USB-access guidance.
+func pickThorRecoveryDevice() (thorDevice, error) {
+	for {
+		devs, err := rcm.ListRecoveryDevices()
+		switch {
+		case errors.Is(err, rcm.ErrUSBAccess) && len(devs) == 0:
+			fmt.Println("\n" + usbAccessHintBox())
+			fmt.Print("Press Enter to rescan, or 'q' to quit: ")
+			if readQuit() {
+				return thorDevice{}, fmt.Errorf("cannot open the Jetson's USB device: permission denied")
+			}
+			continue
+		case errors.Is(err, rcm.ErrUSBAccess):
+			fmt.Println()
+			fmt.Println(tui.WarningMessage("Another Jetson in recovery mode was detected but couldn't be opened (USB access denied) — it is NOT listed below."))
+		case err != nil:
+			return thorDevice{}, err
+		}
+		if len(devs) == 0 {
+			// The user already confirmed the Thor is in recovery mode, so an empty
+			// scan usually means cabling or the button sequence needs another try.
+			// Wait passively (spinner) until a device appears or the user quits.
+			if devs, err = waitForThorRecovery(rcm.ListRecoveryDevices); err != nil {
+				return thorDevice{}, err
+			}
+		}
+		switch len(devs) {
+		case 1:
+			return thorDevice{PathKey: devs[0].PathKey, Label: devs[0].Describe()}, nil
+		default:
+			var items []tui.PickerItem
+			byKey := make(map[string]rcm.RecoveryDevice, len(devs))
+			for _, d := range devs {
+				byKey[d.PathKey] = d
+				items = append(items, tui.PickerItem{
+					Name:    d.Describe(),
+					Section: "Recovery devices",
+					SortKey: d.PathKey,
+					Value:   d.PathKey,
+				})
+			}
+			sel, err := pickFromItems("Select the Thor to flash", items)
+			if err != nil {
+				return thorDevice{}, err
+			}
+			return thorDevice{PathKey: byKey[sel].PathKey, Label: byKey[sel].Describe()}, nil
+		}
+	}
+}
+
+// thorIsUSBAccessErr reports whether err is the OS denying wendy access to the
+// Jetson's USB device (rcm.ErrUSBAccess), which routes into usbAccessHintBox.
+func thorIsUSBAccessErr(err error) bool {
+	return errors.Is(err, rcm.ErrUSBAccess)
+}
+
+// diskAvailBytes reports the bytes available to the user on the volume holding
+// path. ok=false when statfs fails or reports implausible values (buggy FUSE
+// drivers), in which case the disk-space preflight is skipped.
+func diskAvailBytes(path string) (avail int64, ok bool) {
+	var stat unix.Statfs_t
+	if err := unix.Statfs(path, &stat); err != nil {
+		return 0, false
+	}
+	if stat.Bsize <= 0 {
+		return 0, false // implausible block size: can't size the disk safely
+	}
+	avail = int64(stat.Bavail) * int64(stat.Bsize)
+	if avail < 0 {
+		return 0, false
+	}
+	return avail, true
+}
+
+// thorStageOne performs the stage-1 RCM boot over gousb.
+func thorStageOne(fp *flashpack.Flashpack, dev thorDevice, out io.Writer) error {
+	return bringup.Run(bringup.Options{
+		Dir:        fp.Stage1Dir(),
+		MemBCT:     fp.MemBCT(),
+		DevicePath: dev.PathKey,
+		SendOrder:  fp.Manifest.Stage1SendOrder,
+		Out:        out,
+	})
+}
+
+// thorOpenGadget opens the initrd-flash ADB gadget as a stage-2 transport,
+// pinning the selected device via WENDY_ADB_PATH and retrying while it
+// re-enumerates after stage-1.
+func thorOpenGadget(dev thorDevice, out io.Writer) (flashengine.Transport, func(), error) {
+	if dev.PathKey != "" {
+		os.Setenv("WENDY_ADB_PATH", dev.PathKey)
+	}
+	deadline := time.Now().Add(60 * time.Second)
+	var lastErr error
+	for time.Now().Before(deadline) {
+		d, err := adb.Open()
+		if err == nil {
+			return d, func() { d.Close() }, nil
+		}
+		lastErr = err
+		time.Sleep(time.Second)
+	}
+	// Nothing was written yet, so signal the "gadget unreachable, Thor is safe" path.
+	return nil, nil, fmt.Errorf("%w (over ADB): %v", errGadgetUnreachable, lastErr)
+}
