--- go/internal/cli/commands/os_install_thor_other.go
@@ -1,13 +1,0 @@
-//go:build !darwin && !linux
-
-package commands
-
-import (
-	"context"
-	"fmt"
-)
-
-// installThor is macOS/Linux-only (USB recovery flashing uses gousb/libusb).
-func installThor(_ context.Context, _ string, _ bool, _ bool) error {
-	return fmt.Errorf("Thor (jetson-agx-thor) flashing is currently only supported on macOS and Linux")
-}
