+++ go/internal/cli/commands/os_install_thor_test.go
@@ -1,5 +1,3 @@
-//go:build darwin || linux
-
 package commands
 
 import (
